dice moved to utils.js
